Add resource operation authorization
Added resource based authorization for crud operations

+++ src/TobyMeehan.Com.Api/Features/Application/Delete/Endpoint.cs
@@ -1,16 +1,21 @@
 using FastEndpoints;
+using Microsoft.AspNetCore.Authorization;
 using TobyMeehan.Com.Api.Requests;
+using TobyMeehan.Com.Api.Security;
 using TobyMeehan.Com.Services;
+using IAuthorizationService = Microsoft.AspNetCore.Authorization.IAuthorizationService;
 
 namespace TobyMeehan.Com.Api.Features.Application.Delete;
 
 public class Endpoint : Endpoint<IdRequest>
 {
     private readonly IApplicationService _service;
+    private readonly IAuthorizationService _authorizationService;
 
-    public Endpoint(IApplicationService service)
+    public Endpoint(IApplicationService service, IAuthorizationService authorizationService)
     {
         _service = service;
+        _authorizationService = authorizationService;
     }
     
     public override void Configure()
@@ -21,10 +26,25 @@ public class Endpoint : Endpoint<IdRequest>
 
     public override async Task HandleAsync(IdRequest req, CancellationToken ct)
     {
-        var result = await _service.DeleteAsync(new Id<IApplication>(req.Id), ct);
+        var result = await _service.GetByIdAsync(new Id<IApplication>(req.Id), cancellationToken: ct);
 
-        await result.Match(
-            success => SendNoContentAsync(ct),
-            notFound => SendNotFoundAsync(ct));
+        if (!result.IsSuccess(out var application))
+        {
+            await SendNotFoundAsync(ct);
+            return;
+        }
+        
+        var authorizationResult = 
+            await _authorizationService.AuthorizeAsync(User, application, OperationRequirements.Delete);
+
+        if (!authorizationResult.Succeeded)
+        {
+            await SendForbiddenAsync(ct);
+            return;
+        }
+        
+        await _service.DeleteAsync(new Id<IApplication>(req.Id), ct);
+
+        await SendNoContentAsync(ct);
     }
 }
